+++ NEWS.md
@@ -2,3 +2,5 @@
 
 * Initial version, implementing Cheysson color palettes, patterns and fonts
 * Fixed problem with fonts, requiring `showtext::showtext_auto()`
+* Added Getting started vignette
+* Added Guerry maps vignette